+++ frontend/src/buttons/LaunchButton/LaunchButton.tsx
@@ -21,7 +21,7 @@ import {
 } from '@material-ui/core'
 import { Icon } from '../../components/Icon'
 import { emit } from '../../services/Controller'
-import { ModalSetUsername } from '../../components/ModalSetUsername'
+import { UsernameModal } from '../../components/UsernameModal'
 
 type Props = {
   connection?: IConnection
@@ -39,18 +39,13 @@ export const LaunchButton: React.FC<Props> = ({ connection, service, menuItem, s
   const css = useStyles()
   const { ui } = useDispatch<Dispatch>()
   const app = useApplication(service && service.typeID)
-  const [openUsername, setOpenUsername] = useState<boolean>(false)
-  const [username, setUsername] = useState<string>((connection && connection.username) || '')
+  const [open, setOpen] = useState<boolean>(false)
   const [openPutty, setOpenPutty] = useState<boolean>(false)
   const closeAll = () => {
-    setOpenUsername(false)
+    setOpen(false)
     setOpenPutty(false)
     ui.set({ requireInstallPutty: false })
   }
-
-  useEffect(() => {
-    setUsername(connection?.username || '')
-  }, [connection?.username])
 
   useEffect(() => {
     requireInstallPutty ? setOpenPutty(true) : closeAll()
@@ -60,16 +55,16 @@ export const LaunchButton: React.FC<Props> = ({ connection, service, menuItem, s
 
   const check = () => {
     if (!app.prompt || connection.username) launch()
-    else setOpenUsername(true)
+    else setOpen(true)
   }
-  const launch = () => {
+  const launch = (username?: string) => {
     closeAll()
     if (username)
       setConnection({
         ...connection,
         username: username.toString(),
       })
-    const launchApp = app.launch({ ...connection, username })
+    const launchApp = app.launch({ ...connection, username: username || connection.username})
 
     launchBrowser(app.title) ? window.open(launchApp) : emit('service/launch', { command: launchApp, pathPutty })
   }
@@ -103,15 +98,7 @@ export const LaunchButton: React.FC<Props> = ({ connection, service, menuItem, s
         </Tooltip>
       )}
 
-      <ModalSetUsername
-        connection={connection}
-        openUsername={openUsername}
-        handleSubmit={launch}
-        username={username}
-        setUsername={setUsername}
-        service={service}
-        close={closeAll}
-      />
+      <UsernameModal connection={connection} open={open} onSubmit={launch} service={service} onClose={closeAll} />
 
       <Dialog open={openPutty} onClose={closeAll} maxWidth="xs" fullWidth>
         <Typography variant="h1">Please install Putty to launch SSH connections.</Typography>
