--- frontend/src/components/ModalSetUsername/index.ts
@@ -1,1 +1,0 @@
-export { ModalSetUsername } from './ModalSetUsername'
