+++ frontend/src/components/UsernameModal/index.ts
@@ -1,0 +1,1 @@
+export { UsernameModal } from './UsernameModal'
